Added some super sick parameter checking

+++ clean_java.sh
@@ -4,7 +4,12 @@
 # Basically the gist is that sed isn't the same on mac os vs linux
 # So we have to test the versioning to get the right behavior
 sedi () {
-    sed --version >/dev/null 2>&1 && sed -i -- "$@" || sed -i "" "$@"
+  sed --version >/dev/null 2>&1 && sed -i -- "$@" || sed -i "" "$@"
+}
+
+
+usage() {
+  echo "usage: $0 <SOURCE FILE> <DESTINATION FILE> [NOISE PARENT DIRECTORY]"
 }
 
 # This is where we remove a bunch of gunk in the java files
@@ -12,7 +17,7 @@ NOISE_PATTERNS="noise.txt"
 LINE_NOISE="line_noise.txt"
 
 if [[ "$#" -lt 2 ]]; then
-  echo "usage: $0 <SOURCE FILE> <DESTINATION FILE>"
+  usage
 	exit 1
 fi
 
@@ -25,6 +30,25 @@ if [[ "$#" -eq 3 ]]; then
   NOISE_PATTERNS="$PARENT_DIR/$NOISE_PATTERNS" # Kind of ghetto navigate back but w/e
   LINE_NOISE="$PARENT_DIR/$LINE_NOISE"
 fi
+
+# Make assertions on them args!
+if [[ ! -f $FILE ]]; then
+  echo "Could not locate source file: [$FILE], exiting"
+  usage
+  exit 1
+fi
+
+if [[ ! -f $NOISE_PATTERNS ]]; then
+  echo "Could not locate noise file: [$NOISE_PATTERNS], exiting"
+  usage
+  exit 1
+fi
+
+if [[ ! -f $LINE_NOISE ]]; then
+  echo ">>Warning: Could not locate line noise file: [$LINE_NOISE]
+  This file greatly improves the utility of cleaned files"
+fi
+
 
 # Make a backup of the input file to work on
 cp $FILE $DESTINATION_FILE
@@ -49,8 +73,10 @@ tr '[:upper:]' '[:lower:]' < $DESTINATION_FILE > $TEMP; mv $TEMP $DESTINATION_FI
 # Now we do the reverse grep on the silly patterns
 grep -i -v -f $NOISE_PATTERNS $DESTINATION_FILE > $TEMP; mv $TEMP $DESTINATION_FILE
 
-# Remove the lame lines as exact matches
-grep -x -v -F -f $LINE_NOISE $DESTINATION_FILE > $TEMP; mv $TEMP $DESTINATION_FILE
+# Remove the lame lines as exact matches, if file exists
+if [[ -f $LINE_NOISE ]]; then
+  grep -x -v -F -f $LINE_NOISE $DESTINATION_FILE > $TEMP; mv $TEMP $DESTINATION_FILE
+fi
 
 # Finally sort things to make our lives easier
 sort $DESTINATION_FILE | sed "s/[[:blank:]]*$//" | uniq > $TEMP;  mv $TEMP $DESTINATION_FILE
